requirements.txt
Django
django-allauth
django-filter
django-widget-tweaks
mysqlclient
python-decouple
yagmail
